refactor: [automl] seperate automl reduction vs automl impl (#4006)
* refactor: [automl] split io model to seperate file

* create directory for automl

* split file to util tbd

* more shuffling

* cleanup includes

* update champs score and add missing metrics
