+++ test/test-sets/ref/aml_ccb_metrics.json
@@ -1,1 +1,1 @@
-{"cbea_label_first_action":1,"cbea_label_not_first":3,"cbea_labeled_ex":4,"cbea_max_actions":2,"cbea_min_actions":2,"cbea_non_zero_cost":0,"cbea_predict_in_learn":0,"conf_idx_0":0,"conf_idx_1":6,"conf_idx_2":5,"current_champ":0,"sfm_count_learn_example_with_shared":4,"test_county":4,"total_champ_switches":0,"total_learn_calls":4,"total_log_calls":0,"total_predict_calls":0,"upcnt_0":0,"upcnt_1":3,"upcnt_2":3,"bound_0":0.0,"bound_1":0.0,"bound_2":0.0,"cbea_avg_actions_per_event":2.0,"cbea_avg_feat_per_action":0.0,"cbea_avg_feat_per_event":0.0,"cbea_avg_ns_per_action":0.0,"cbea_avg_ns_per_event":0.0,"cbea_sum_cost":0.0,"cbea_sum_cost_baseline":0.0,"ips_0":0.0,"ips_1":0.0,"ips_2":0.0,"r_0":0.0,"r_1":-0.0,"r_2":-0.0,"w_0":0.0,"w_1":0.0,"w_2":0.0,"exclusionc_0":"{}","exclusionc_1":"{[\"U\", \"U\"]}","exclusionc_2":"{[\"S\", \"U\"]}","interactions_0":"-q AA -q AS -q AU -q SS -q SU -q UU ","interactions_1":"-q AA -q AS -q AU -q SS -q SU ","interactions_2":"-q AA -q AS -q AU -q SS -q UU "}
+{"cbea_label_first_action":1,"cbea_label_not_first":3,"cbea_labeled_ex":4,"cbea_max_actions":2,"cbea_min_actions":2,"cbea_non_zero_cost":0,"cbea_predict_in_learn":0,"conf_idx_amls_0":0,"conf_idx_amls_1":6,"conf_idx_amls_2":5,"current_champ":0,"sfm_count_learn_example_with_shared":4,"test_county":4,"total_champ_switches":0,"total_learn_calls":4,"total_log_calls":0,"total_predict_calls":0,"upcnt_amls_0":4,"upcnt_amls_1":3,"upcnt_amls_2":3,"upcnt_sc_0":4,"upcnt_sc_1":3,"upcnt_sc_2":3,"bound_amls_0":0.0,"bound_amls_1":0.0,"bound_amls_2":0.0,"bound_sc_0":0.0,"bound_sc_1":0.0,"bound_sc_2":0.0,"cbea_avg_actions_per_event":2.0,"cbea_avg_feat_per_action":0.0,"cbea_avg_feat_per_event":0.0,"cbea_avg_ns_per_action":0.0,"cbea_avg_ns_per_event":0.0,"cbea_sum_cost":0.0,"cbea_sum_cost_baseline":0.0,"ips_amls_0":0.0,"ips_amls_1":0.0,"ips_amls_2":0.0,"ips_sc_0":0.0,"ips_sc_1":0.0,"ips_sc_2":0.0,"r_amls_0":-0.0,"r_amls_1":-0.0,"r_amls_2":-0.0,"r_sc_0":-0.0,"r_sc_1":-0.0,"r_sc_2":-0.0,"w_amls_0":0.0,"w_amls_1":0.0,"w_amls_2":0.0,"w_sc_0":0.0,"w_sc_1":0.0,"w_sc_2":0.0,"exclusionc_0":"{}","exclusionc_1":"{[\"U\", \"U\"]}","exclusionc_2":"{[\"S\", \"U\"]}","interactions_amls_0":"-q AA -q AS -q AU -q SS -q SU -q UU ","interactions_amls_1":"-q AA -q AS -q AU -q SS -q SU ","interactions_amls_2":"-q AA -q AS -q AU -q SS -q UU "}
