+++ test/test-sets/ref/metrics_cbadfautoml.json
@@ -1,1 +1,1 @@
-{"conf_idx_0":0,"conf_idx_1":3,"conf_idx_2":2,"current_champ":0,"line_parse_error":0,"number_events_zero_actions":0,"number_skipped_events":0,"sfm_count_learn_example_with_shared":3,"test_county":3,"total_champ_switches":0,"total_learn_calls":3,"total_log_calls":1,"total_predict_calls":0,"upcnt_0":0,"upcnt_1":2,"upcnt_2":2,"bound_0":0.0,"bound_1":0.0,"bound_2":0.0,"dsjson_sum_cost_original":0.0,"dsjson_sum_cost_original_baseline":0.0,"ips_0":0.0,"ips_1":0.0,"ips_2":0.0,"r_0":0.0,"r_1":1.0,"r_2":1.0,"w_0":0.0,"w_1":0.0,"w_2":0.0,"exclusionc_0":"{}","exclusionc_1":"{[\"E\", \"G\"]}","exclusionc_2":"{[\"E\", \"M\"]}","first_event_id":"0074434d3a3a46529f65de8a59631939","first_event_time":"","interactions_0":"-q EE -q EG -q EM -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","interactions_1":"-q EE -q EM -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","interactions_2":"-q EE -q EG -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","last_event_id":"9077f996581148978a0ebe2484260dab","last_event_time":""}
+{"conf_idx_amls_0":0,"conf_idx_amls_1":3,"conf_idx_amls_2":2,"current_champ":0,"line_parse_error":0,"number_events_zero_actions":0,"number_skipped_events":0,"sfm_count_learn_example_with_shared":3,"test_county":3,"total_champ_switches":0,"total_learn_calls":3,"total_log_calls":1,"total_predict_calls":0,"upcnt_amls_0":3,"upcnt_amls_1":2,"upcnt_amls_2":2,"upcnt_sc_0":3,"upcnt_sc_1":2,"upcnt_sc_2":2,"bound_amls_0":0.0,"bound_amls_1":0.0,"bound_amls_2":0.0,"bound_sc_0":0.0,"bound_sc_1":0.0,"bound_sc_2":0.0,"dsjson_sum_cost_original":0.0,"dsjson_sum_cost_original_baseline":0.0,"ips_amls_0":0.0,"ips_amls_1":0.0,"ips_amls_2":0.0,"ips_sc_0":0.0,"ips_sc_1":0.0,"ips_sc_2":0.0,"r_amls_0":1.0,"r_amls_1":1.0,"r_amls_2":1.0,"r_sc_0":1.0,"r_sc_1":1.0,"r_sc_2":1.0,"w_amls_0":0.0,"w_amls_1":0.0,"w_amls_2":0.0,"w_sc_0":0.0,"w_sc_1":0.0,"w_sc_2":0.0,"exclusionc_0":"{}","exclusionc_1":"{[\"E\", \"G\"]}","exclusionc_2":"{[\"E\", \"M\"]}","first_event_id":"0074434d3a3a46529f65de8a59631939","first_event_time":"","interactions_amls_0":"-q EE -q EG -q EM -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","interactions_amls_1":"-q EE -q EM -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","interactions_amls_2":"-q EE -q EG -q EO -q ER -q ES -q ET -q Ei -q Ej -q GG -q GM -q GO -q GR -q GS -q GT -q Gi -q Gj -q MM -q MO -q MR -q MS -q MT -q Mi -q Mj -q OO -q OR -q OS -q OT -q Oi -q Oj -q RR -q RS -q RT -q Ri -q Rj -q SS -q ST -q Si -q Sj -q TT -q Ti -q Tj -q ii -q ij -q jj ","last_event_id":"9077f996581148978a0ebe2484260dab","last_event_time":""}
