+++ test/train-sets/ref/automl_readable.txt
@@ -98,7 +98,7 @@ _automl_config_manager_configs[10]_state = New
 _automl_config_manager_scores.size() = 3
 _automl_config_manager_scores[0].first_chisq_alpha = 0.05000000074505806
 _automl_config_manager_scores[0].first_chisq_tau = 1
-_automl_config_manager_scores[0].first_chisq_n = 0
+_automl_config_manager_scores[0].first_chisq_n = 3
 _automl_config_manager_scores[0].first_chisq_sumw = 0
 _automl_config_manager_scores[0].first_chisq_sumwsq = 0
 _automl_config_manager_scores[0].first_chisq_sumwr = 0
@@ -109,14 +109,14 @@ _automl_config_manager_scores[0].first_chisq_rmax = 1
 _automl_config_manager_scores[0].first_chisq_wmin = 0
 _automl_config_manager_scores[0].first_chisq_wmax = inf
 _automl_config_manager_scores[0].first_ips = 0
-_automl_config_manager_scores[0].first_count = 0
+_automl_config_manager_scores[0].first_count = 3
 _automl_config_manager_scores[0].first_lastw = 0
-_automl_config_manager_scores[0].first_lastr = 0
+_automl_config_manager_scores[0].first_lastr = 1
 _automl_config_manager_scores[0].first_index = 0
 _automl_config_manager_scores[0].first_eligible_to_inactivate = false
 _automl_config_manager_scores[0].second_chisq_alpha = 0.05000000074505806
 _automl_config_manager_scores[0].second_chisq_tau = 1
-_automl_config_manager_scores[0].second_chisq_n = 0
+_automl_config_manager_scores[0].second_chisq_n = 3
 _automl_config_manager_scores[0].second_chisq_sumw = 0
 _automl_config_manager_scores[0].second_chisq_sumwsq = 0
 _automl_config_manager_scores[0].second_chisq_sumwr = 0
@@ -127,9 +127,9 @@ _automl_config_manager_scores[0].second_chisq_rmax = 1
 _automl_config_manager_scores[0].second_chisq_wmin = 0
 _automl_config_manager_scores[0].second_chisq_wmax = inf
 _automl_config_manager_scores[0].second_ips = 0
-_automl_config_manager_scores[0].second_count = 0
+_automl_config_manager_scores[0].second_count = 3
 _automl_config_manager_scores[0].second_lastw = 0
-_automl_config_manager_scores[0].second_lastr = 0
+_automl_config_manager_scores[0].second_lastr = 1
 _automl_config_manager_scores[1].first_chisq_alpha = 0.05000000074505806
 _automl_config_manager_scores[1].first_chisq_tau = 1
 _automl_config_manager_scores[1].first_chisq_n = 2
